requirements.txt
Django~=3.2.23
pyshorteners~=1.0.1
djangorestframework~=3.14.0